changed to use decimal instead of double

+++ backend/Models/ProbabilityResponse.cs
@@ -2,6 +2,6 @@ namespace RedingtonCalculator.Models
 {
     public class ProbabilityResponse
     {
-        public double CalculatedValue { get; set; }
+        public decimal CalculatedValue { get; set; }
     }
 }
